Gamerfeeds/api/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
newsapi-python==0.2.7
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
